SERVER-43762 tighten the overload set for BSONObjBuilder::appendNumber
GitOrigin-RevId: 98f42065a3f530024980e72ef0a15a371c28e32b

+++ src/mongo/db/modules/enterprise/src/fle/commands/cryptd_ismaster.cpp
@@ -58,7 +58,7 @@ public:
 
         result.appendBool("iscryptd", true);
         result.appendNumber("maxBsonObjectSize", BSONObjMaxUserSize);
-        result.appendNumber("maxMessageSizeBytes", MaxMessageSizeBytes);
+        result.appendNumber("maxMessageSizeBytes", static_cast<long long>(MaxMessageSizeBytes));
         result.appendDate("localTime", jsTime());
 
         // Mongos tries to keep exactly the same version range of the server for which
